HBX-1899: Rename package 'org.hibernate.mvn' to 'org.hibernate.tool.maven'
